Replaced routers directory and added controllers.
