+++ src/router.js
@@ -1,0 +1,8 @@
+'use strict';
+
+const orders = require('./controllers/orders');
+
+module.exports = (app) => {
+    app.get('/api/orders', orders.list);
+    app.post('/api/order', orders.create);
+};
